--- app/com/robocubs4205/cubscout/controllers/HomeController.scala
@@ -1,25 +1,0 @@
-package com.robocubs4205.cubscout.controllers
-
-import com.robocubs4205.cubscout._
-import javax.inject._
-import play.api._
-import play.api.mvc._
-
-/**
- * This controller creates an `Action` to handle HTTP requests to the
- * application's home page.
- */
-@Singleton
-class HomeController @Inject()(cc: ControllerComponents) extends AbstractController(cc) {
-
-  /**
-   * Create an Action to render an HTML page.
-   *
-   * The configuration in the `routes` file means that this method
-   * will be called when the application receives a `GET` request with
-   * a path of `/`.
-   */
-  def index() = Action { implicit request: Request[AnyContent] =>
-    Ok(views.html.index())
-  }
-}
