--- app/com/robocubs4205/cubscout/views/index.scala.html
@@ -1,5 +1,0 @@
-@()
-
-@main("Welcome to Play") {
-  <h1>Welcome to Play!</h1>
-}
