+++ conf/routes
@@ -3,31 +3,28 @@
 # https://www.playframework.com/documentation/latest/ScalaRouting
 # ~~~~
 
-# An example controller showing a sample home page
-GET     /                           com.robocubs4205.cubscout.controllers.HomeController.index
-
 # Map static resources from the /public folder to the /assets URL path
-GET     /assets/*file               controllers.Assets.versioned(path="/public", file: Asset)
+GET     /assets/*file                   controllers.Assets.versioned(path="/public", file: Asset)
 
-GET     /v1/districts               com.robocubs4205.cubscout.controllers.apiv1.DistrictController.index(context:Option[String])
-POST    /v1/districts               com.robocubs4205.cubscout.controllers.apiv1.DistrictController.create(context:Option[String])
-GET     /v1/districts/:id           com.robocubs4205.cubscout.controllers.apiv1.DistrictController.get(id:Long,context:Option[String])
-DELETE  /v1/districts/:id           com.robocubs4205.cubscout.controllers.apiv1.DistrictController.delete(id:Long,context:Option[String])
-PUT     /v1/districts/:id           com.robocubs4205.cubscout.controllers.apiv1.DistrictController.put(id:Long,context:Option[String])
-GET     /v1/districts/:id/events    com.robocubs4205.cubscout.controllers.apiv1.DistrictController.eventsIndex(id:Long,context:Option[String])
-GET     /v1/districts/:id/teams     com.robocubs4205.cubscout.controllers.apiv1.DistrictController.teamsIndex(id:Long,context:Option[String])
+GET     /api/v1/districts               com.robocubs4205.cubscout.api.v1.controllers.DistrictController.index(context:Option[String])
+POST    /api/v1/districts               com.robocubs4205.cubscout.api.v1.controllers.DistrictController.create(context:Option[String])
+GET     /api/v1/districts/:id           com.robocubs4205.cubscout.api.v1.controllers.DistrictController.get(id:Long,context:Option[String])
+DELETE  /api/v1/districts/:id           com.robocubs4205.cubscout.api.v1.controllers.DistrictController.delete(id:Long,context:Option[String])
+PUT     /api/v1/districts/:id           com.robocubs4205.cubscout.api.v1.controllers.DistrictController.put(id:Long,context:Option[String])
+GET     /api/v1/districts/:id/events    com.robocubs4205.cubscout.api.v1.controllers.DistrictController.eventsIndex(id:Long,context:Option[String])
+GET     /api/v1/districts/:id/teams     com.robocubs4205.cubscout.api.v1.controllers.DistrictController.teamsIndex(id:Long,context:Option[String])
 
-GET     /v1/events                  com.robocubs4205.cubscout.controllers.apiv1.EventController.index(context:Option[String])
-POST    /v1/events                  com.robocubs4205.cubscout.controllers.apiv1.EventController.create(context:Option[String])
-GET     /v1/events/:id              com.robocubs4205.cubscout.controllers.apiv1.EventController.get(id:Long,context:Option[String])
-DELETE  /v1/events/:id              com.robocubs4205.cubscout.controllers.apiv1.EventController.delete(id:Long,context:Option[String])
-PUT     /v1/events/:id              com.robocubs4205.cubscout.controllers.apiv1.EventController.put(id:Long,context:Option[String])
-GET     /v1/events/:id/matches      com.robocubs4205.cubscout.controllers.apiv1.EventController.matchIndex(id:Long,context:Option[String])
+GET     /api/v1/events                  com.robocubs4205.cubscout.api.v1.controllers.EventController.index(context:Option[String])
+POST    /api/v1/events                  com.robocubs4205.cubscout.api.v1.controllers.EventController.create(context:Option[String])
+GET     /api/v1/events/:id              com.robocubs4205.cubscout.api.v1.controllers.EventController.get(id:Long,context:Option[String])
+DELETE  /api/v1/events/:id              com.robocubs4205.cubscout.api.v1.controllers.EventController.delete(id:Long,context:Option[String])
+PUT     /api/v1/events/:id              com.robocubs4205.cubscout.api.v1.controllers.EventController.put(id:Long,context:Option[String])
+GET     /api/v1/events/:id/matches      com.robocubs4205.cubscout.api.v1.controllers.EventController.matchIndex(id:Long,context:Option[String])
 
-GET     /v1/games                   com.robocubs4205.cubscout.controllers.apiv1.GameController.index(context:Option[String])
-POST    /v1/games                   com.robocubs4205.cubscout.controllers.apiv1.GameController.create(context:Option[String])
-GET     /v1/games/:id               com.robocubs4205.cubscout.controllers.apiv1.GameController.get(id:Long,context:Option[String])
-DELETE  /v1/games/:id               com.robocubs4205.cubscout.controllers.apiv1.GameController.delete(id:Long,context:Option[String])
-PUT     /v1/games/:id               com.robocubs4205.cubscout.controllers.apiv1.GameController.replace(id:Long,context:Option[String])
-GET     /v1/games/:id/events        com.robocubs4205.cubscout.controllers.apiv1.GameController.eventIndex(id:Long,context:Option[String])
-GET     /v1/games/:id/robots        com.robocubs4205.cubscout.controllers.apiv1.GameController.robotIndex(id:Long,context:Option[String])
+GET     /api/v1/games                   com.robocubs4205.cubscout.api.v1.controllers.GameController.index(context:Option[String])
+POST    /api/v1/games                   com.robocubs4205.cubscout.api.v1.controllers.GameController.create(context:Option[String])
+GET     /api/v1/games/:id               com.robocubs4205.cubscout.api.v1.controllers.GameController.get(id:Long,context:Option[String])
+DELETE  /api/v1/games/:id               com.robocubs4205.cubscout.api.v1.controllers.GameController.delete(id:Long,context:Option[String])
+PUT     /api/v1/games/:id               com.robocubs4205.cubscout.api.v1.controllers.GameController.replace(id:Long,context:Option[String])
+GET     /api/v1/games/:id/events        com.robocubs4205.cubscout.api.v1.controllers.GameController.eventIndex(id:Long,context:Option[String])
+GET     /api/v1/games/:id/robots        com.robocubs4205.cubscout.api.v1.controllers.GameController.robotIndex(id:Long,context:Option[String])
